fix(Place) : #238 - Place KakaoId 타입 수정
int -> long

Related to: #238

+++ backend/enjoy-service/src/main/java/com/enjoyservice/api/course/dto/CourseListRes.java
@@ -46,7 +46,7 @@ public class CourseListRes {
             @JsonProperty("place_id")
             private long placeId;
             @JsonProperty("kakao_id")
-            private int kakaoId;
+            private long kakaoId;
             @JsonProperty("place_name")
             private String placeName;
             @JsonProperty("img_url")
